# api-docs.yml
openapi: 3.0.0
...
info:
  title: amoxtli
  version: "1.0"
servers:
  - url: "http://localhost:3000"
paths:
...
  /libros:
    get:
      summary: Recuperar libros
      tags: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaLibros"
      operationId: getLibros
      x-stoplight:
        id: t7zlzjxlrwh5i
      description: Recupera una lista de todos los libros ofrecidos por usuarios
      security: []
      parameters:
        - schema:
            type: integer
          in: query
          name: pagina
          description: Página a recuperar
  "/libros/{isbn}":
    parameters:
      - schema:
          type: string
        name: isbn
        in: path
        required: true
        description: ISBN del libro
    get:
      summary: Recuperar libro
      tags: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibroConDuenos"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getLibro
      x-stoplight:
        id: 0dmmns2snfs0g
      description: Recupera un la información de un libro
      security: []
  "/libros/{isbn}/detalles":
    parameters:
      - schema:
          type: string
        name: isbn
        in: path
        required: true
        description: ISBN del libro
    get:
      summary: Recuperar información de un libro
      tags: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetallesLibro"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      operationId: getDetallesLibro
      x-stoplight:
        id: h94tofvp3rnxd
      description: Recupera información detallada de un libro mediante su ISBN
      security: []
...
components:
  schemas:
...
    Error:
      x-stoplight:
        id: sdbjvvqfbhfq9
      type: object
      description: Información de un error
      properties:
        mensaje:
          type: string
          x-stoplight:
            id: nyb49827nm7xq
          description: Mensaje amigable que describe el error
        codigo:
          type: integer
          x-stoplight:
            id: on1qw5ccbym94
          description: Código único asociado con el error
      required:
        - mensaje
        - codigo
...
    InfoBasicaLibro:
      type: object
      x-stoplight:
        id: 9edfaf04d6bfe
      x-examples: {}
      description: Información básica de un libro para ser mostrado en una lista
      properties:
        isbn:
          type: string
          description: ISBN (International Standard Book Number) del libro
          x-faker: phonphoneNumberei
        autor:
          type: string
          description: Autor del libro
          x-faker: name.findName
        titulo:
          type: string
          description: Título del libro
          x-faker:
            random.words:
              - 4
          x-stoplight:
            id: h44grtsuocgq6
        urlPortada:
          type: string
          format: uri
          description: URL de la portada del libro
          x-faker:
            fake:
              - "http://loremflickr.com/480/640/abstract"
      required:
        - isbn
        - autor
        - titulo
        - urlPortada
...
    DetallesLibro:
      x-examples:
        Example 1:
          isbn: string
          titulo: string
          autor: string
          genero: string
          editorial: string
          sinopsis: string
          foto: string
          id: 0
      description: Información detallada respecto a un libro
      allOf:
        - $ref: "#/components/schemas/InfoBasicaLibro"
        - type: object
          properties:
            generos:
              type: array
              description: Géneros literarios del libro
              items:
                x-stoplight:
                  id: jlfqp7nmrxuom
                type: string
                x-faker: random.word
            editorial:
              type: string
              description: Editorial que publica el libro
              x-faker: company.companyName
            sinopsis:
              type: string
              description: Sinopsis del libro
              x-faker: lorem.paragraphs
            idioma:
              type: string
              x-stoplight:
                id: vpapnbvg25med
              description: Idioma del libro
              x-faker: random.locale
            fechaPublicacion:
              type: string
              x-stoplight:
                id: 6zbf0op55imig
              format: date
              description: Fecha de publicación del libro
              x-faker:
                date.past:
                  - 100
          required:
            - generos
            - editorial
            - idioma
            - fechaPublicacion
    LibroConDuenos:
      x-examples: {}
      allOf:
        - $ref: "#/components/schemas/InfoBasicaLibro"
        - $ref: "#/components/schemas/DetallesLibro"
        - type: object
          properties:
            duenos:
              x-stoplight:
                id: ozskd6akltet9
              type: array
              description: Lista de usuarios que tienen este libro en su catálogo
              items:
                $ref: "#/components/schemas/Dueno"
          required:
            - duenos
      description: "Información detallada de un libro, incluyendo "
...
    Dueno:
      type: object
      x-examples: {}
      description: Detalles de un usuario que es dueño de un libro
      properties:
        id:
          type: integer
          description: ID del usuario
        nombre:
          type: string
          description: Nombre del usuario
          x-faker: name.firstName
        apellido:
          type: string
          description: Apellido o apellidos del usuario
          x-faker: name.lastName
        foto:
          type: string
          format: uri
          description: Foto de perfil del usuario
          x-faker: image.avatar
          nullable: true
        idLibro:
          type: integer
          x-stoplight:
            id: 9up4y3vqpb7kf
      required:
        - id
        - nombre
        - apellido
        - foto
        - idLibro
...
    PaginaLibros:
      title: PaginaLibros
      x-stoplight:
        id: q0ub4iloeo26w
      type: object
      properties:
        pagSiguiente:
          type: integer
          x-stoplight:
            id: nvwzfe4vyz3o2
          description: Número de página siguiente
        pagAnterior:
          type: integer
          x-stoplight:
            id: nqw4j6vsn12vz
          description: Número de página anterior
        resultados:
          type: integer
          x-stoplight:
            id: aasvjid0x0cze
          description: Número de resultados por página
        libros:
          type: array
          items:
            $ref: "#/components/schemas/InfoBasicaLibro"
      required:
        - pagSiguiente
        - pagAnterior
        - resultados
        - libros
...
  securitySchemes:
    Token:
      type: http
      scheme: bearer
      description: ""
security:
  - Token: []